+++ QuanLySieuThi/src/main/java/com/example/quanlysieuthi/data/entity/PhieuNhap.java
@@ -17,7 +17,7 @@ public class PhieuNhap {
     private String ma;
     private Date ngayLapPhieuNhap;
     private Integer soLuong;
-    private Integer tongTien;
+    private Long tongTien;
 
     @ManyToOne
     @JoinColumn(name = "ma_nhan_vien")
